organise existing components in atoms/molecules/organisms
