+++ client/src/components/atoms/index.js
@@ -1,0 +1,4 @@
+export { AppHeading } from './AppHeading';
+export { Button } from './Button';
+export { Box } from './Box';
+export { Container } from './Container';
