+++ client/src/components/index.js
@@ -1,5 +1,2 @@
-export { AppHeading } from './AppHeading';
-export { Box } from './Box';
-export { Button } from './Button';
-export { Container } from './Container';
-export { ItemCard } from './ItemCard';
+export { Box, Button, AppHeading, Container } from './atoms';
+export { ItemCard } from './molecules';
