+++ client/src/components/molecules/ItemCard/ItemCard.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import styled from 'styled-components';
 import { Card as AntCard } from 'antd';
+import { Card } from './ItemCard.styles';
 
 const { Meta } = AntCard;
 
@@ -25,23 +25,3 @@ export const ItemCard = ({ item }) => {
     </Card>
   );
 };
-
-const Card = styled(AntCard)`
-  flex: 1;
-  max-width: 13rem;
-  min-width: 13rem;
-  margin: 1rem;
-
-  @media (max-width: 768px) {
-    margin: 0.5rem 0.3rem;
-    max-width: 9rem;
-    min-width: 9rem;
-
-    .ant-card-meta-title {
-      font-size: 14px;
-    }
-    .ant-card-meta-description {
-      font-size: 10px;
-    }
-  }
-`;
