+++ client/src/components/molecules/ItemCard/ItemCard.styles.js
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+import { Card as AntCard } from 'antd';
+
+export const Card = styled(AntCard)`
+  flex: 1;
+  max-width: 13rem;
+  min-width: 13rem;
+  margin: 1rem;
+
+  @media (max-width: 768px) {
+    margin: 0.5rem 0.3rem;
+    max-width: 9rem;
+    min-width: 9rem;
+
+    .ant-card-meta-title {
+      font-size: 0.8rem;
+    }
+    .ant-card-meta-description {
+      font-size: 0.6rem;
+    }
+  }
+`;
