+++ client/src/components/molecules/ItemCard/index.js
@@ -1,0 +1,1 @@
+export { ItemCard } from './ItemCard';
